Added organization by category and disclaimer messages
The program now organizes files based on their category
For example, .mp3, .wav, and .m4a files are all placed into a single 'Audio' folder instead of separate folders based on their file extensions.

There are currently four categories: Audio, Video, Image, and Document.

If the mimetype returns None by python, the 'categorize_by_type' function classifies it as 'Unknown'. In these cases, the file is placed in a folder based on its extension.
For example, blender files '.blend' is currently returned as 'Unknown', so the program will put the .blend files into a folder named 'blend' accroding to their file extension.

Additionally, the program now displays a disclaimer showing the user the path they are going organize, the file types found inside the directory given and the categories the files are going to be sorted into